requirements/devel.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
packaging==25.0
Pygments==2.19.1
requests==2.32.4
setuptools==80.9.0
urllib3==2.4.0
Werkzeug==3.1.3
PyJWT==2.8.0
